Initial file upload
Current structure is HTML page supported by styling, functions, and config file. Config file stores bulk constants for now. Due to API limits for free QuickBase account, we are not yet storing these parameters in the back end. This page is also intended to be copied. The config file requires one less step in this process.